Remove the `drop_with_repr_extern` lint.
This lint was removed in nightly rust.  Note that while this lint still
exists in the Stable and Beta channels, it only applies if we have types
with `#[repr(C)]` that implement `Drop`.  Since this library has neither
(and will probably never have them), it is safe to remove this lint.