Breaking API change:  encode and decode bytes, not strings.
This commit makes an API change to the `encode`, `decode`, and
`MultiHash` functions/types, to operate on bytes, not strings.

In rust, all strings are valid UTF-8, and it's impossible to construct
one that's not.  With the previous API, only UTF-8 data could be hashed,
so the new API operates on bytes instead.

The other API change is for encode to return a Vec<u8> of bytes, instead
of a hex-encoded string.  The hex-encoded string was not incorrect, but
wasn't a nice as it could be, as it forced this particular encoding on
the user.

Other changes:
 * No longer export the `hashes` module, since its contents are already
   exported.
 * Removed the rustc_serialized dependency.
 * Made `to_hex` public, as it might be a useful utility function (and
   removed some unnecessary calls to `.rev()`

Since this is a Breaking Change, the version number has been bumped