Rename `encode` to `sum`, introduce new `encode`
In go-multihash, the `Sum` function takes un-hashed data and a
specification of a hash function and produces multihash bytes
by hashing the input data. The `Encode` function, on the other hand,
produces multihash bytes from an existing hash digest.

This commit adds an equivalent of the `Encode` function from
go-multihash. Since there is already a function named "encode",
this commit also renames it to `sum` to make it similar to the go
version of this library. This is a breaking change.

Closes #25