feat: remove code table (#1)
Instead of using a code table, derive from a Multihash enum which defines
implementations.

Also split the `MultihashDigest` trait into the parts that need a hasher implementation
and those who don't. This way Multihashes can be created that don't have a specific
hasher associated, this is what `RawMultihash` is for.