feat: rename MultihashCode to MultihashDigest
While I was looking into upgrading libp2p to this version of multihash,
I realized that `MultihashCode` might not be such a good name for the
trait.

First of all, it's kind of a meaningless name. What does it actually do?
It does create digests, hence I think `MultihashDigest` is a better name.

Also `MultihashCode` can easily clash with other names. For example if
you have a custom Multihash code table, just calling it `Code` might be
a to generic name in case you want to have it as a top-level export. Then
`MultihashCode` might be a good name, but it would clash with the trait
name.

BREAKING CHANGE: The `MultihashCode` trait is now named `MultihashDigest`.