feat: create Multihash that matches the code table
Instead of only export the `Multihash` that is generic over the allocated
size, also expose `Multihash` that is created by the `Multihash` derive.

If the `multihash-impl` feature is enabled, also a `Multihash` with the
corresponding size gets exported.

BREAKING CHANGE: The `Multihash` which is generic over the size is now
called `MultihashGeneric`.