fix: remove the "identity" feature
There are _always_ safer ways to compute an identity hash, and this
feature can cause otherwise "safe" code to crash. With this feature
enabled, `Code::try_from(codec)?.digest(input)` would crash if `codec`
was 0 and `input` was more than 64 bytes.

Unfortunately, this was a cargo "feature" so a single dependency
enabling it means it's enabled for all other crates in the build.

Users of this code migrate to `Multihash::wrap(0, digest)`, which
returns an error if the digest is too large instead of panicing.

fixes #194