fix(codetable): automatically include digest dep when needed
Previously it was needed to manually enable the optional digest
dependency when using any of sha1,2,3, or ripemd features, and
it could only be found out via reading the sources or codetable
failed to build.

Also included some additional feature guards to clean up warnings
or errors that ocured when building without all features.